% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanFun.R
\name{MeanAppt}
\alias{MeanAppt}
\title{Mean by Column}
\usage{
MeanAppt(x)
}
\arguments{
\item{x}{"Data Frame you are pulling from" $ "column you are pulling from"}
}
\value{
The average number of that appointment status on that given day of the week
}
\description{
Mean by Column
}
\examples{
MeanFun(ApptData$MonA)

}
