% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercWalkInFun.R
\name{PercWalkIn}
\alias{PercWalkIn}
\title{Percent Walk-In Function}
\usage{
PercWalkIn(x, y)
}
\arguments{
\item{x}{is the number of appointment walk-ins on any given day}

\item{y}{is the total number of vehicles seen on any given day}
}
\value{
The percentage of walk-ins we had on that given day
}
\description{
Percent Walk-In Function
}
\examples{
PercWalkIn(13,33)

}
